Move Config out of the client. Distinguish between Openshift and Kubernetes config.
