+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/OpenshiftConfig.java
@@ -1,0 +1,65 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.fabric8.kubernetes.client;
+
+import io.fabric8.kubernetes.client.internal.Utils;
+import io.sundr.builder.annotations.Buildable;
+
+public class OpenshiftConfig extends Config {
+
+  private String oapiVersion = "v1";
+  private String openShiftUrl;
+  private Config kubernetesConfig;
+
+  public OpenshiftConfig() {
+    this(new ConfigBuilder().build());
+  }
+
+  public OpenshiftConfig(Config kubernetesConfig) {
+    this(kubernetesConfig, "v1", null);
+  }
+
+  @Buildable(builderPackage = "io.fabric8.kubernetes.api.builder")
+  public OpenshiftConfig(Config kubernetesConfig, String oapiVersion, String openShiftUrl) {
+    super(kubernetesConfig.isTrustCerts(),
+      kubernetesConfig.getMasterUrl(), kubernetesConfig.getApiVersion(), kubernetesConfig.getEnabledProtocols(),
+      kubernetesConfig.getCaCertFile(), kubernetesConfig.getCaCertData(),
+      kubernetesConfig.getClientCertFile(), kubernetesConfig.getClientCertData(),
+      kubernetesConfig.getClientKeyFile(), kubernetesConfig.getClientKeyData(),
+      kubernetesConfig.getClientKeyAlgo(), kubernetesConfig.getClientKeyPassphrase(),
+      kubernetesConfig.getUsername(), kubernetesConfig.getPassword(), kubernetesConfig.getOauthToken());
+    this.kubernetesConfig=kubernetesConfig;
+    this.oapiVersion = Utils.getSystemPropertyOrEnvVar(KUBERNETES_OAPI_VERSION_SYSTEM_PROPERTY, oapiVersion);
+    this.openShiftUrl = openShiftUrl;
+
+    if (this.openShiftUrl == null) {
+      this.openShiftUrl = getMasterUrl() + "oapi/" + this.oapiVersion + "/";
+    }
+  }
+
+  public Config getKubernetesConfig() {
+    return this.kubernetesConfig;
+  }
+
+  public String getOapiVersion() {
+    return oapiVersion;
+  }
+
+  public String getOpenShiftUrl() {
+    return openShiftUrl;
+  }
+}
