+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/examples/FullExample.java
@@ -17,6 +17,8 @@ package io.fabric8.kubernetes.client.examples;
 
 import com.ning.http.client.ws.WebSocket;
 import io.fabric8.kubernetes.api.model.*;
+import io.fabric8.kubernetes.client.*;
+import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
@@ -35,7 +37,7 @@ public class FullExample {
       master = args[0];
     }
 
-    DefaultKubernetesClient.Config config = new DefaultKubernetesClient.ConfigBuilder().masterUrl(master).build();
+    Config config = new ConfigBuilder().withMasterUrl(master).build();
     try (final KubernetesClient client = new DefaultKubernetesClient(config)) {
       try {
         // Create a namespace for all our stuff
@@ -102,12 +104,7 @@ public class FullExample {
             .endSpec().done());
 
         // Get the RC by name in namespace
-        ReplicationController gotRc = client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").get();
-        log("Get RC by name in namespace", gotRc);
-        // Dump the RC as YAML
-        log("Dump RC as YAML", SerializationUtils.dumpAsYaml(gotRc));
-        log("Dump RC as YAML without state", SerializationUtils.dumpWithoutRuntimeStateAsYaml(gotRc));
-
+        log("Get RC by name in namespace", client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").get());
         // Get the RC by label
         log("Get RC by label", client.replicationControllers().withLabel("server", "nginx").list());
         // Get the RC without label
@@ -120,9 +117,6 @@ public class FullExample {
         log("Get RC by label in namespace", client.replicationControllers().inNamespace("thisisatest").withLabel("server", "nginx").list());
         // Update the RC
         client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").edit().editMetadata().addToLabels("new", "label").endMetadata().done();
-
-        Thread.sleep(1000);
-
         // Update the RC - change the image to apache
         client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").edit().editSpec().editTemplate().withNewSpec()
           .addNewContainer().withName("nginx").withImage("httpd")
@@ -131,11 +125,6 @@ public class FullExample {
           .endSpec()
           .endTemplate()
           .endSpec().done();
-
-        Thread.sleep(1000);
-
-        // Update the RC - change the image back to nginx using a rolling update
-        client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").rollImageUpdate("nginx");
 
         Thread.sleep(1000);
 
@@ -152,7 +141,7 @@ public class FullExample {
         client.replicationControllers().inNamespace("thisisatest").withName("nginx2-controller").delete();
         log("Deleted RCs");
 
-        //Create another RC inline
+        //Create an ohter RC inline
         client.replicationControllers().inNamespace("thisisatest").createNew().withNewMetadata().withName("nginx-controller").addToLabels("server", "nginx").endMetadata()
           .withNewSpec().withReplicas(3)
           .withNewTemplate()
@@ -165,8 +154,6 @@ public class FullExample {
           .endTemplate()
           .endSpec().done();
          log("Created inline RC");
-
-        Thread.sleep(1000);
 
         client.replicationControllers().inNamespace("thisisatest").withName("nginx-controller").delete();
         log("Deleted RC");
